add JavaDoc, make text fields use ValidationController
