+++ vaadin-text-field-flow-parent/vaadin-text-field-flow/src/main/java/com/vaadin/flow/component/textfield/AbstractNumberField.java
@@ -24,6 +24,7 @@ import java.util.function.Function;
 import com.vaadin.flow.component.AttachEvent;
 import com.vaadin.flow.component.Synchronize;
 import com.vaadin.flow.component.shared.ClientValidationUtil;
+import com.vaadin.flow.component.shared.ValidationController;
 import com.vaadin.flow.component.shared.ValidationUtil;
 import com.vaadin.flow.data.binder.ValidationResult;
 import com.vaadin.flow.data.binder.ValidationStatusChangeEvent;
@@ -57,14 +58,57 @@ public abstract class AbstractNumberField<C extends AbstractNumberField<C, T>, T
     private boolean stepSetByUser;
     private boolean minSetByUser;
 
-    private boolean manualValidationEnabled = false;
-
     private DomListenerRegistration inputListenerRegistration;
 
     private final CopyOnWriteArrayList<ValidationStatusChangeListener<T>> validationStatusChangeListeners = new CopyOnWriteArrayList<>();
 
-    private String customErrorMessage;
-    private String constraintErrorMessage;
+    private Validator<T> defaultValidator = (value, context) -> {
+        boolean hasBadInput = valueEquals(value, getEmptyValue())
+                && isInputValuePresent();
+        if (hasBadInput) {
+            return ValidationResult.error(getI18nErrorMessage(
+                    AbstractNumberFieldI18n::getBadInputErrorMessage));
+        }
+
+        if (context == null) {
+            ValidationResult requiredResult = ValidationUtil
+                    .validateRequiredConstraint(getI18nErrorMessage(
+                            AbstractNumberFieldI18n::getRequiredErrorMessage),
+                            isRequiredIndicatorVisible(), value,
+                            getEmptyValue());
+            if (requiredResult.isError()) {
+                return requiredResult;
+            }
+        }
+
+        Double doubleValue = value != null ? value.doubleValue() : null;
+
+        ValidationResult maxResult = ValidationUtil.validateMaxConstraint(
+                getI18nErrorMessage(
+                        AbstractNumberFieldI18n::getMaxErrorMessage),
+                doubleValue, max);
+        if (maxResult.isError()) {
+            return maxResult;
+        }
+
+        ValidationResult minResult = ValidationUtil.validateMinConstraint(
+                getI18nErrorMessage(
+                        AbstractNumberFieldI18n::getMinErrorMessage),
+                doubleValue, min);
+        if (minResult.isError()) {
+            return minResult;
+        }
+
+        if (!isValidByStep(value)) {
+            return ValidationResult.error(getI18nErrorMessage(
+                    AbstractNumberFieldI18n::getStepErrorMessage));
+        }
+
+        return ValidationResult.ok();
+    };
+
+    private ValidationController<AbstractNumberField<C, T>, T> validationController = new ValidationController<>(
+            this);
 
     /**
      * Sets up the common logic for number fields.
@@ -118,8 +162,7 @@ public abstract class AbstractNumberField<C extends AbstractNumberField<C, T>, T
      */
     @Override
     public void setErrorMessage(String errorMessage) {
-        customErrorMessage = errorMessage;
-        updateErrorMessage();
+        validationController.setCustomErrorMessage(errorMessage);
     }
 
     /**
@@ -129,20 +172,7 @@ public abstract class AbstractNumberField<C extends AbstractNumberField<C, T>, T
      */
     @Override
     public String getErrorMessage() {
-        return customErrorMessage;
-    }
-
-    private void setConstraintErrorMessage(String errorMessage) {
-        constraintErrorMessage = errorMessage;
-        updateErrorMessage();
-    }
-
-    private void updateErrorMessage() {
-        String errorMessage = constraintErrorMessage;
-        if (customErrorMessage != null && !customErrorMessage.isEmpty()) {
-            errorMessage = customErrorMessage;
-        }
-        getElement().setProperty("errorMessage", errorMessage);
+        return validationController.getCustomErrorMessage();
     }
 
     @Override
@@ -344,7 +374,7 @@ public abstract class AbstractNumberField<C extends AbstractNumberField<C, T>, T
 
     @Override
     public Validator<T> getDefaultValidator() {
-        return (value, context) -> checkValidity(value, false);
+        return defaultValidator;
     }
 
     @Override
@@ -367,55 +397,9 @@ public abstract class AbstractNumberField<C extends AbstractNumberField<C, T>, T
                 .forEach(listener -> listener.validationStatusChanged(event));
     }
 
-    private ValidationResult checkValidity(T value,
-            boolean withRequiredValidator) {
-        boolean hasBadInput = valueEquals(value, getEmptyValue())
-                && isInputValuePresent();
-        if (hasBadInput) {
-            return ValidationResult.error(getI18nErrorMessage(
-                    AbstractNumberFieldI18n::getBadInputErrorMessage));
-        }
-
-        if (withRequiredValidator) {
-            ValidationResult requiredResult = ValidationUtil
-                    .validateRequiredConstraint(getI18nErrorMessage(
-                            AbstractNumberFieldI18n::getRequiredErrorMessage),
-                            isRequiredIndicatorVisible(), value,
-                            getEmptyValue());
-            if (requiredResult.isError()) {
-                return requiredResult;
-            }
-        }
-
-        Double doubleValue = value != null ? value.doubleValue() : null;
-
-        ValidationResult maxResult = ValidationUtil.validateMaxConstraint(
-                getI18nErrorMessage(
-                        AbstractNumberFieldI18n::getMaxErrorMessage),
-                doubleValue, max);
-        if (maxResult.isError()) {
-            return maxResult;
-        }
-
-        ValidationResult minResult = ValidationUtil.validateMinConstraint(
-                getI18nErrorMessage(
-                        AbstractNumberFieldI18n::getMinErrorMessage),
-                doubleValue, min);
-        if (minResult.isError()) {
-            return minResult;
-        }
-
-        if (!isValidByStep(value)) {
-            return ValidationResult.error(getI18nErrorMessage(
-                    AbstractNumberFieldI18n::getStepErrorMessage));
-        }
-
-        return ValidationResult.ok();
-    }
-
     @Override
     public void setManualValidation(boolean enabled) {
-        this.manualValidationEnabled = enabled;
+        validationController.setManualValidation(enabled);
     }
 
     /**
@@ -428,18 +412,7 @@ public abstract class AbstractNumberField<C extends AbstractNumberField<C, T>, T
      * The method does nothing if the manual validation mode is enabled.
      */
     protected void validate() {
-        if (this.manualValidationEnabled) {
-            return;
-        }
-
-        ValidationResult result = checkValidity(getValue(), true);
-        if (result.isError()) {
-            setInvalid(true);
-            setConstraintErrorMessage(result.getErrorMessage());
-        } else {
-            setInvalid(false);
-            setConstraintErrorMessage("");
-        }
+        validationController.validate(getValue());
     }
 
     private boolean isValidByStep(T value) {
